Rework Proxy type, replace i32 with an Enum(u32)
The Proxy type which wrapped an i32 now wraps a ProxyState, which
is an enum consisting of PendingAdd, PendingRemove(index), Removed and Added(index)
representing the current state of variable from the POV of the Gurobi library.

The index() method now returns a Result<u32> instead of an i32, and an error
is given when the inner ProxyType does not take a Added(_) value.

The Proxy type cannot be made private, since Constr/Var etc implement
Deref<Target=Proxy>.

The update() method is now correctly called when creating Var or Constr
objects, if lazy updates are disabled.

Remove caching of update mode for model.  The sequence of operations is broken:
1. create model with param::UpdateMode = 0
2. call get_update_mode() (indirectly)
3. change param::UpdateMode to 1
4. call get_update_mode() (indirectly) incorrectly returns 0

Bugfix in add_constrs and add_ranges: LinExpr objects were not being converted into
CSR format properly

Into<(parts)> impls have been replaced with an into_parts() method in the type impl
for QuadExpr and LinExpr types.

Tests have been refined, but more are still needed.

This is large breaking change, more will probably come.