Rework Proxy type
The Proxy type, which was an Rc, Cell-wrapped IndexState has been removed

Index states will be tracked instead using the IdxManager, which is essentially HashMap
 that maps ModelObject => IdxState and a Vec<ModelObject> to track the ordering of these
 objects, with respect to the order in which they were first added to the model.

A ModelObject trait has been added which is directly implemented by Var, Constr, QConstr
and SOS.  This bundles all the behaviours requried for interacting with IdxManager objects.

The advantage to this approach is that the state for each object is managed in one place,
and it allows for some optimisations.  Most importantly to users, since ModelObjects now
just contain an ID and Model ID, they now implement copy which makes building expressions
much more ergonomic.

Several bugs around the eager update behaviour have been fixed in the process.  Namely,
when objects are added to a Model with eager updates, they DON'T immediately get attributes
from Gurobi.  As a result, there are now 4 IdxStates that IdxManager has to worry about:

Pending - An object has been added to the model, but cannot be interacted with or used for building.
    This happens when the model is in lazy-update mode (UpdateMode = 0).
Build - An object has been added to the model, can be used to building and setting attributes,
    but not querying attributes.  This happens when the model is in eager-update mode (UpdateMode = 1)
Present - An object is present in the model and can be fully used. This is the state of a Pending or Build
    variable after a model update.
Removed - An object has been removed from the model, but its index is still tracked by Gurobi.  After
    an update these objects can be fully dropped from the IdxManager.

Remove itertools dependency.

Add macros for adding variables.