Fix GRBModel outliving GRBEnv
In the Gurobi C API, there seem to be two kinds of *GRBEnv pointers:

User-Allocated (UA): These are obtain through calls to GRBemptyenv and GRBloadenv,
and must be freed using GRBfreenv AFTER any GRBModel objects have been freed.

Gurobi-Allocated (GA): These are created and freed by Gurobi internally, and
should not be freed using GRBfreeenv, as this causes a double-free.

The previous implementation of Env stored a *GRBEnv and a bool indicated whether
the pointer was UA or GA.  This prevented the double-free.  However,
it did NOT guarantee, as per the Gurobi docs, that a UA GRBEnv used to construct
a GRBModel would outlive that model.

For example, this sequence of operations broke the spec:
create Env (alloc GRBEnv)
create Model (alloc GRBModel)
drop Env - possibly due to scoping (free GRBEnv)
drop Model (free GRBModel)

Therefore, it is necessary keep track of the original UA GRBEnv that each Model refers to.

The reason for the two *GRBEnv types is unknown, but can be speculated as follows based
on the Gurobi docs, which state that creating an environment is expensive.

Creating a UA GRBEnv parses and checks a license file, sets up cluster connections and does other
network IO.  This information is mostly static.

When creating a GA GRBEnv, this static information points to an existing UA GRBEnv.
The mutable part (from the POV of the user) is things like model parameters.  When
coping a model (GRBcopymodel), or creating a fixed or feasbility relaxation
(GRBfixmodel/GRBfeasrelax) version of a model, a GA GRBEnv is created. Because
it points back to the original UA GRBEnv, the model cannot outlive the UA GRBEnv.

Solution:
The UA GRBEnv gets its own type, UserAllocEnv, which stores the pointer.  When this type
is Drop'ed, GRBfreeenv is called.

The Env struct now stores a Rc<UserAllocEnv>, to ensure the UA environment is kept for
as long as necessary.  If Env represents a GA GRBEnv, it stores also a *mut GRBEnv.
Thus, Env has a second field which is an Option<*mut GRBEnv>, which is None iff
Env represents a UA GRBEnv.

As a result, when creating a Model from a raw pointer, such as in Model::copy, the original
Model's environment is required, so we can Clone the Rc<UserAllocEnv>.