Rework callback API
Keeping references in closures can get extremely annoying.  Given
that most use cases of callbacks will keep some sort of mutable state,
it is much more ergonmic to keep this state in a struct and instead
call a trait method on the struct.

The Callback trait has been added which has a single required method
that takes a mutable self-ref and the callback context object (now called CbCtx).

This allows users easier access to their own datastructures, without having to wrap
everything in Rc<RefCell<_>>.  The trait has a blanket impl on FnMut(CbCtx) -> Result<()>
so closures may still be used if the user prefers.

This is a breaking change, since the Model::optimize_with_callback method now takes
a reference to a Callback object, rather than an owned FnMut.  Closures objects therefore
must be declared as mutable (even though they won't be mutated) and passed as a reference.