Rework callback context object
Previously, the callback context was a single type which contained all possible values.
This change defines a callback context type for each possible "WHERE" value.  This allows
us to restrict the methods on each context available to only those allowed by the Gurobi API.
For example, calling set_solution in PreSolve would be an error.

As a result, the context object has been moved inside the Where enum, to allow the user to
obtain the particular context type using a match pattern.

Aside from reducing user errors, the attributes such as NodeCount have been moved to getter-methods,
to avoid fetching them eagerly (which in most cases is not needed).

Additionally, the Callback trait has been changed to return an anyhow::Result<()> from the
callback() method.  This allows users to return errors from other libraries from their callback.
A type alias has been added to make the purpose of this result type clear.  This adds anyhow as
a dependency.