Implement GurobiName more selectively
AttrName and ParamName are now merged into GurobiName.

The default impl used unsafe code, which may lead to bugs down the line.  Unlikely, as although GurobiName is public,
it is not importable and therefore not implementable.  Nonetheless, a marker trait (GurobiNameMarker) has been added to all
attr and param enums instead.  This means a blanket impl can be used which only targets the attr/param enums.