Rework AttachModel trait
AttachModel only allows the user to supply a Model to the `attach() method`.  The trait has been re-worked into AttachVarName
which accepts allows the user to supply anything that implements QueryVarName (a new help trait). This way, a Model or a
HashMap<Var, String> can be supplied for more flexibility.  The old AttachModel trait is kept for stability, but is
now deprecated.

The main use-case for this is for pretty-printing `Expr`/`IneqExpr` in contexts where we don't have access to a Model reference,
such as within a callback (since `Model::optimize_with_callback` takes &mut self).